requirements.txt
streamlink
fastapi
opencv-python
uvicorn[standard]
ffmpeg-python
ultralytics
salesforce-lavis
openai-whisper
pydub